+++ saos-persistence/src/main/java/pl/edu/icm/saos/persistence/repository/RawSourceJudgmentRepository.java
@@ -14,10 +14,10 @@ import pl.edu.icm.saos.persistence.model.importer.RawSourceJudgment;
 /**
  * @author madryk
  */
-public interface RawSourceJudgmentRepository extends JpaRepository<RawSourceJudgment, Integer> {
+public interface RawSourceJudgmentRepository extends JpaRepository<RawSourceJudgment, Long> {
   
     @Query("select rJudgment from RawSourceJudgment rJudgment where TYPE(rJudgment)=:clazz and rJudgment.id=:id")
-    <T extends RawSourceJudgment> T findOne(@Param("id") Integer id, @Param("clazz") Class<T> clazz);
+    <T extends RawSourceJudgment> T findOne(@Param("id") Long id, @Param("clazz") Class<T> clazz);
     
     
     @Query("select rJudgment from RawSourceJudgment rJudgment where TYPE(rJudgment)=:clazz and rJudgment.sourceId=:sourceId")
@@ -29,7 +29,7 @@ public interface RawSourceJudgmentRepository extends JpaRepository<RawSourceJudg
 
     
     @Query("select rJudgment.id from RawSourceJudgment rJudgment where TYPE(rJudgment) = :clazz and rJudgment.processed=false order by id")
-    List<Integer> findAllNotProcessedIds(@Param("clazz") Class<? extends RawSourceJudgment> clazz);
+    List<Long> findAllNotProcessedIds(@Param("clazz") Class<? extends RawSourceJudgment> clazz);
   
     
     @Transactional
