+++ saos-persistence/src/main/java/pl/edu/icm/saos/persistence/repository/ScChamberRepositoryCustomImpl.java
@@ -1,13 +1,13 @@
 package pl.edu.icm.saos.persistence.repository;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import pl.edu.icm.saos.persistence.common.InitializingVisitor;
-import pl.edu.icm.saos.persistence.model.CommonCourt;
-import pl.edu.icm.saos.persistence.model.SupremeCourtChamber;
-
 import javax.persistence.EntityManager;
 import javax.transaction.Transactional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import pl.edu.icm.saos.persistence.common.InitializingVisitor;
+import pl.edu.icm.saos.persistence.model.SupremeCourtChamber;
 
 /**
  * @author pavtel
@@ -22,7 +22,7 @@ public class ScChamberRepositoryCustomImpl implements ScChamberRepositoryCustom 
 
     @Override
     @Transactional
-    public SupremeCourtChamber findOneAndInitialize(int id) {
+    public SupremeCourtChamber findOneAndInitialize(long id) {
         SupremeCourtChamber chamber = entityManager.find(SupremeCourtChamber.class, id);
 
         if(chamber != null){
